requirements.txt
streamlit
torch
torchvision
pillow
numpy
lime
matplotlib
scikit-image